+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 - 2020-11-10
+
+### Added
+- Route and route group compiler and factory.
+
+### Removed
+- Support for old way of registering routes and route groups.
+- Support for authorization registration, moved to direct services.
+
 ## 1.0.1 - 2020-08-10
 
 ### Fixed
@@ -15,5 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The initial implementation of the package.
 
 # Versions
-- [1.0.1 > Unreleased](https://github.com/ulrack/web/compare/1.0.1...HEAD)
+- [2.0.0 > Unreleased](https://github.com/ulrack/web/compare/2.0.0...HEAD)
+- [1.0.1 > 2.0.0](https://github.com/ulrack/web/compare/1.0.1...2.0.0)
 - [1.0.0 > 1.0.1](https://github.com/ulrack/web/compare/1.0.0...1.0.1)